requirements.txt
pyserial==3.5
pyserial-asyncio==0.6
aiohttp==3.9.5